backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
charset-normalizer==3.3.2
Django==3.2.25
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
idna==3.7
mysqlclient==2.1.1
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
